9068: Moved filter list order check to more appropriate test
